+++ modules/attendancetracker/attendanceTrackerTask.ts
@@ -2,8 +2,8 @@ import {Client, TextChannel} from "discord.js";
 import { get } from "../../lib/configmanager";
 import { Weekday } from "../../lib/recurringtask";
 import {TaskExecutor} from "../../types";
-import {buildAttendanceAction, buildTimeTableEmbed} from "./attendancetrackerVisuals";
-import {getBlocks, resetAttendance} from "./attendancetracker";
+import {buildAttendanceAction, buildTimeTableEmbed} from "./attendanceTrackerVisuals";
+import {getBlocks, resetAttendance} from "./attendanceTracker";
 
 /**
  * TaskExecutor for printing the timetable with attendance buttons of the given weekday <br>
