+++ modules/truthtable/truthtableParser.ts
